Hotfix: 엔티티, dto 맴버변수 타입 변경 (#18)
* feat: entity, dto 객체 멤버변수 타입 변환
- entity의 아이디 타입을 Long으로 변환
- entity,dto 멤버변수 타입을 wrapper class로 변환

* test: entity, dto 객체 멤버변수 타입 변환에 따른 테스트 코드 수

+++ src/main/java/com/addiction/alertHistory/repository/AlertHistoryJpaRepository.java
@@ -8,7 +8,7 @@ import com.addiction.alertHistory.entity.AlertHistory;
 
 public interface AlertHistoryJpaRepository extends JpaRepository<AlertHistory, Long> {
 
-	void deleteByUserId(int userId);
+	void deleteByUserId(Long userId);
 
-	List<AlertHistory> findByUserId(int userId);
+	List<AlertHistory> findByUserId(Long userId);
 }
