+++ src/main/java/com/addiction/survey/surveyAnswer/repository/SurveyAnswerJpaRepository.java
@@ -5,8 +5,8 @@ import org.springframework.data.jpa.repository.JpaRepository;
 import com.addiction.survey.surveyAnswer.entity.SurveyAnswer;
 import com.addiction.survey.surveyQuestion.entity.SurveyQuestion;
 
-public interface SurveyAnswerJpaRepository extends JpaRepository<SurveyAnswer, Integer> {
+public interface SurveyAnswerJpaRepository extends JpaRepository<SurveyAnswer, Long> {
 
-	int findScoreById(int id);
+	Integer findScoreById(Long id);
 
 }
