+++ src/main/java/com/addiction/user/users/service/request/UserUpdateSurveyServiceRequest.java
@@ -9,14 +9,14 @@ import lombok.Getter;
 
 @Getter
 public class UserUpdateSurveyServiceRequest {
-	private final List<Integer> answerId;
+	private final List<Long> answerId;
 	private final String purpose;
 	private final int cigarettePrice;
 	private final Sex sex;
 	private final String birthDay;
 
 	@Builder
-	public UserUpdateSurveyServiceRequest(List<Integer> answerId, String purpose, int cigarettePrice, Sex sex,
+	public UserUpdateSurveyServiceRequest(List<Long> answerId, String purpose, int cigarettePrice, Sex sex,
 		String birthDay) {
 		this.answerId = answerId;
 		this.purpose = purpose;
